Rework the build, CI and release lifecycle
meson-python replaces setuptools as the build backend, and make becomes
the single front end for it. CI pipelines completely redesigned (see
.github/PIPELINES.md). The tree gains REUSE 3.3 compliance, a pinned
ruff rule set, clang-tidy, and a pre-commit configuration that defines
linting.

+++ modules/fhe/python/test/test_gp25.py
@@ -1,3 +1,4 @@
+# SPDX-FileCopyrightText: 2026 Antonio Guimarães <antonio.guimaraes@imdea.org>
 # SPDX-License-Identifier: Apache-2.0
 """Characterization tests for the (reverted) vfhe.fhe GP25 sparse-amortized
 bootstrap over the cffi boundary.
@@ -33,7 +34,7 @@ def get_min_prec(key):
                 r_max = max(r_max, previous - (N - j - 1))
                 previous = N - j - 1
         r_max = max(r_max, previous)
-    return int(math.floor(math.log2(r_max))) + 1
+    return math.floor(math.log2(r_max)) + 1
 
 
 def rs_sparse_ternary_key(scheme, h, sigma, target_r_prec):
